am65xx: Fix Compilation warnings on int to pointer conversion

+++ packages/ti/sdo/ipc/family/am65xx/InterruptHost.c
@@ -48,8 +48,8 @@
 #include "package/internal/InterruptHost.xdc.h"
 
 /* Register access method. */
-#define REG16(A)   (*(volatile UInt16 *) (A))
-#define REG32(A)   (*(volatile UInt32 *) (A))
+#define REG16(A)   (*(volatile UInt16 *) ((uintptr_t)A))
+#define REG32(A)   (*(volatile UInt32 *) ((uintptr_t)A))
 
 #define PROCID(IDX)               (InterruptHost_procIdTable[(IDX)])
 #define MBX_TABLE_IDX(SRC, DST)   ((PROCID(SRC) * InterruptHost_NUM_CORES) + \
